renamed count to numCompletedAppointments
added numCancellations

removed buggy amPm to id string
