+++ imports/api/stats.ts
@@ -13,10 +13,20 @@ export interface Stats {
   totalDuration: number; // service duration
   estimatedWaitTime: number; // given by system when the queue entry is created
   actualWaitTime: number; // calculated when the service is started
+  numCancellations: number;
+}
 
-  // TODO
-  // num_cancellations: number;
-  // total_wait_time
+export interface StatsData {
+  serviceId: string;
+  date: Date;
+  inc?: {
+    numCompletedAppointments?: number;
+    startTime?: Date;
+    endTime?: Date;
+    estimatedWaitTime?: number;
+    actualWaitTime?: number;
+    numCancellations?: number;
+  };
 }
 
 export const StatsCollection = new Mongo.Collection<Stats>("stats");
